cast usec to make gcc happy

+++ can2udp/src/can2lan.c
@@ -80,7 +80,7 @@ char *time_stamp(){
     gettimeofday(&tv, &tz);
     tm = localtime(&tv.tv_sec);
 
-    sprintf(timestamp,"%02d:%02d:%02d.%03d", tm->tm_hour, tm->tm_min, tm->tm_sec, tv.tv_usec/1000);
+    sprintf(timestamp,"%02d:%02d:%02d.%03d", tm->tm_hour, tm->tm_min, tm->tm_sec, (int) tv.tv_usec/1000);
     return timestamp;
 }
